requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21
Pillow>=9.0
wandb>=0.15
